// src/app/shared/custom-types.ts
export type AutogeneratedProperties = 'id' | 'joinCode';
export type OmitAutogeneratedProperties<T> = Omit<T, AutogeneratedProperties>;

export type GameModel = {
  id?: string;
  name: string;
  turnLength: number;
  turnLengthUnits: TimeUnits;
  playerIds: PlayerModel['id'][];
  hostPlayerId: PlayerModel['id'] | null;
  joinCode?: string;
};

export type PlayerModel = {
  id?: string;
  name: string;
  color: string;
};

export enum TimeUnits {
  Seconds = 's',
  Minutes = 'min',
}

// TODO: delete this once all references are removed
export enum PlayerColors {
  Red = '#F34141',
  Orange = '#F38141',
  Yellow = '#F3CC41',
  Green = '#38D048',
  Blue = '#4188F3',
  Violet = '#B041F3',
  Pink = '#F659CA',
  Cyan = '#41D9FB',
  Brown = '#8E744D',
  White = '#D3D3D3',
  Gray = '#767676',
  Black = '#181818',
}

export enum LocalStorageKeys {
  Players = 'players',
  Game = 'game',
}

export enum Endpoints {
  GAME = 'game',
  PLAYER = 'player',
}

export enum GameTypes {
  Local = 'local',
  Online = 'online',
}